requirements_bip.txt
# BIP Temporal Invariance Experiment - Requirements
# Core
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
tokenizers>=0.14.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
scikit-learn>=1.3.0

# NLP
spacy>=3.6.0
nltk>=3.8.0
sentence-transformers>=2.2.0

# Hebrew/Aramaic support
hebrew-tokenizer>=2.3.0
pyarabic>=0.6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Experiment tracking
wandb>=0.15.0
tensorboard>=2.14.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0.0
jsonlines>=3.1.0
